Rename .vimrc to vimrc
Unix-like systems look by default to `~/.vim/vimrc`. Changing the name
of `.vimrc` to `vimrc` removes the requirement to symlink the `.vimrc`
in the users home directory if this repo is cloned to `~/.vim`.

If desired one can clone this repo to another location and in this case
symlink the entire `.vim` directory to `~/.vim`.